Making the bits of thrift that Liger Uses comply with "-Wshorten-64-32"
Summary: Title says it.

Test Plan: built and ran thrift tests as well as testing on iOS device

Reviewed By: alandau@fb.com

Subscribers: alandau, bmatheny, seanc, mshneer

FB internal diff: D1804811

Signature: t1:1804811:1422388665:7e7b2bbcd78bfd7d7456b2634b719c48e1c0bb08

+++ thrift/lib/cpp/transport/TSSLSocket.cpp
@@ -206,7 +206,7 @@ void TSSLSocket::checkHandshake() {
 void TSSLSocket::verifyCertificate() {
 
   // verify authentication result
-  int rc = SSL_get_verify_result(ssl_);
+  long rc = SSL_get_verify_result(ssl_);
   if (rc != X509_V_OK) {
     throw TSSLException(string("SSL_get_verify_result(), ") +
                         X509_verify_cert_error_string(rc));
